Disable `delete (hs->second) in guiGate's destructor
After debugging it appeared that the klsCollisionObject destructor was crashing
because this line of code deleted the object but didn't remove it from the this's
subobjects. Since the child destructor should only handle its own objects I disabled
this line of code. Now the subobjects are simply removed from the hotspots map and
then deleted when the klsCollision destructor calles `deleteSubObjects`