So here's the deal...
I tried overwriting the master branch with a different branch.
Instead, I got some disgusting mash-up of the two without messages
denoting merge conflicts, so I've copied and pasted the files from
the good version into master. Master, as it was before all of these
changes can still be checked out as version-with-blackboxes