Add installer, see full message
I got an email from prof. dudenhoffer asking for
a good build of cedarlogic with an installer. I'm
taking the oppurtunity to leave cedarlogic in a good
state.

In this commit, I add the installer (barebones) and
remove some old config files.

My plan is to tidy the installer, fix the file-finding
bug that happens when cedarlogic is run from a shortcut,
and make this the master branch.