Guard CaptureMouse in beginDrag
It is possible for beginDrag to be called before endDrag for the
previous drag. Which trips a debug assertion.

This can also be triggered by pressing the right mouse button while
holding the left. I discovered it because the mouse I'm using is on it's
last legs, and seems to send a release event followed very quickly by a
press event when dragging sometimes.