Capture mouse on linux to allow drag and drop
GTK doesn't send a mouse enter event to the canvas like happens on
windows. My original solution to this was to use the wxWidgets drag and
drop methods, but this doesn't allow showing the gate on the canvas
before it has been dropped.

Instead this calls beginDrag when the gate is selected in the palette,
allowing the GUICanvas to receive events.